pyproject.toml
[project]
name = "inteligencia-negocios-producto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "albumentations>=2.0.8",
    "matplotlib>=3.10.3",
    "opencv-python>=4.12.0.88",
    "pillow>=9.5.0",
    "pipx>=1.7.1",
    "pre-commit>=4.2.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.2",
    "scikit-learn>=1.2.2",
    "seaborn>=0.13.2",
    "tensorflow[and-cuda]>=2.16.0",
    "tqdm>=4.67.1",
    "visualkeras>=0.1.4",
]

[tool.ruff]
line-length = 80

[tool.ruff.lint]
select = [
    "F401",  # module imported but unused
    "F403",  # from module import * used; unable to detect undefined names
    "F405",  # name may be undefined, or defined from star imports: module
    "F841",  # local variable name is assigned to but never used
    "E501",  # line too long
    "E231",  # missing whitespace after ','
    "E266",  # too many leading '#' for block comment
    "E302",  # expected 2 blank lines, found 1
    "E501",  # line too long
]
